Exclude partials in the `templateselect` default filter. Move methods to `FieldExtension`
