+++ src/Twig/FieldExtension.php
@@ -4,11 +4,18 @@ declare(strict_types=1);
 
 namespace Bolt\Twig;
 
+use Bolt\Common\Str;
 use Bolt\Configuration\Config;
 use Bolt\Entity\Content;
 use Bolt\Entity\Field;
+use Bolt\Entity\Field\SelectField;
+use Bolt\Entity\Field\TemplateselectField;
 use Bolt\Repository\ContentRepository;
 use Bolt\Repository\FieldRepository;
+use Bolt\Storage\Query;
+use Bolt\Utils\ContentHelper;
+use Symfony\Component\Finder\Finder;
+use Symfony\Component\Finder\SplFileInfo;
 use Tightenco\Collect\Support\Collection;
 use Twig\Environment;
 use Twig\Extension\AbstractExtension;
@@ -26,11 +33,24 @@ class FieldExtension extends AbstractExtension
     /** @var Config */
     private $config;
 
-    public function __construct(Notifications $notifications, ContentRepository $contentRepository, Config $config)
+    /** @var ContentHelper */
+    private $contentHelper;
+    
+    /** @var Query */
+    private $query;
+
+    public function __construct(
+        Notifications $notifications,
+        ContentRepository $contentRepository,
+        Config $config,
+        ContentHelper $contentHelper,
+        Query $query)
     {
         $this->notifications = $notifications;
         $this->contentRepository = $contentRepository;
         $this->config = $config;
+        $this->contentHelper = $contentHelper;
+        $this->query = $query;
     }
 
     /**
@@ -50,6 +70,8 @@ class FieldExtension extends AbstractExtension
     {
         return [
             new TwigFunction('field_factory', [$this, 'fieldFactory']),
+            new TwigFunction('list_templates', [$this, 'getListTemplates']),
+            new TwigFunction('select_options', [$this, 'selectOptions']),
         ];
     }
 
@@ -118,5 +140,154 @@ class FieldExtension extends AbstractExtension
         }
 
         return $records;
+    }
+
+    public function getListTemplates(TemplateselectField $field): Collection
+    {
+        $definition = $field->getDefinition();
+        $current = current($field->getValue());
+
+        $finder = new Finder();
+        $templatesDir = $this->config->get('theme/template_directory');
+        $templatesPath = $this->config->getPath('theme', true, $templatesDir);
+
+        $filter = $definition->get('filter', '/^[^_]*\.twig$/');
+
+        if (! Str::isValidRegex($filter)) {
+            $filter = Str::isValidRegex('/' . $filter . '/') ? '/' . $filter . '/' : '/^[^_]*\.twig$/';
+        }
+
+        $finder
+            ->files()
+            ->in($templatesPath)
+            ->path($definition->get('path'))
+            ->sortByName()
+            ->filter(function (SplFileInfo $file) use ($filter) {
+                return preg_match($filter, $file->getRelativePathname()) === 1;
+            });
+
+        $options = [];
+
+        if ($definition->get('required') === false) {
+            $options = [[
+                'key' => '',
+                'value' => '(choose a template)',
+                'selected' => false,
+            ]];
+        }
+
+        foreach ($finder as $file) {
+            $options[] = [
+                'key' => $file->getRelativePathname(),
+                'value' => $file->getRelativePathname(),
+            ];
+
+            if ($current === $file->getRelativePathname()) {
+                $current = false;
+            }
+        }
+
+        if ($current !== false) {
+            $options[] = [
+                'key' => $current,
+                'value' => $current . ' (file seems to be missing)',
+            ];
+        }
+
+        return new Collection($options);
+    }
+
+    public function selectOptions(Field $field): Collection
+    {
+        if (! $field instanceof SelectField) {
+            return collect([]);
+        }
+
+        $values = $field->getOptions();
+
+        if (is_iterable($values)) {
+            return $this->selectOptionsArray($field);
+        }
+
+        return $this->selectOptionsContentType($field);
+    }
+
+    private function selectOptionsArray(Field $field): Collection
+    {
+        if (! $field instanceof SelectField) {
+            return collect([]);
+        }
+
+        $values = $field->getOptions();
+        $currentValues = $field->getValue();
+
+        $options = [];
+
+        // We need to add this as a 'dummy' option for when the user is allowed
+        // not to pick an option. This is needed, because otherwise the `select`
+        // would default to the one.
+        if (! $field->getDefinition()->get('required', true)) {
+            $options[] = [
+                'key' => '',
+                'value' => '',
+                'selected' => false,
+            ];
+        }
+
+        if (! is_iterable($values)) {
+            return new Collection($options);
+        }
+
+        foreach ($values as $key => $value) {
+            $options[] = [
+                'key' => $key,
+                'value' => $value,
+                'selected' => in_array($key, $currentValues, true),
+            ];
+        }
+
+        return new Collection($options);
+    }
+
+    private function selectOptionsContentType(Field $field): Collection
+    {
+        [ $contentTypeSlug, $format ] = explode('/', $field->getDefinition()->get('values'));
+
+        $options = [];
+
+        if (! $field->getDefinition()->get('required')) {
+            $options[] = [
+                'key' => '',
+                'value' => '',
+            ];
+        }
+
+        if (! empty($field->getDefinition()->get('limit'))) {
+            $maxAmount = $field->getDefinition()->get('limit');
+        } else {
+            $maxAmount = $this->config->get('general/maximum_listing_select', 200);
+        }
+
+        $order = $field->getDefinition()->get('order', '');
+
+        $params = [
+            'limit' => $maxAmount,
+            'order' => $order,
+        ];
+
+        /** @var Content[] $records */
+        $records = iterator_to_array($this->query->getContent($contentTypeSlug, $params)->getCurrentPageResults());
+
+        foreach ($records as $record) {
+            if ($field->getDefinition()->get('mode') === 'format') {
+                $formattedKey = $this->contentHelper->get($record, $field->getDefinition()->get('format'));
+            }
+            $options[] = [
+                'key' => $formattedKey ?? $record->getId(),
+                'value' => $this->contentHelper->get($record, $format),
+            ];
+        }
+
+        return new Collection($options);
     }
 }
